Moved things around to remove unnecessary files
