+++ .gitignore
@@ -158,4 +158,4 @@ build/**/*.o
 test.py
 test_nbody.py
 /cli
-/tests/tools/tests_separated/generated_scripts/
+/tests/tools/generated_scripts/*
