+++ tests/tools/body_str.py
